+++ lib/routing/router.dart
@@ -1,17 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
+import 'package:gruene_app/routing/routes.dart';
 import 'package:gruene_app/screens/news/news_screen.dart';
 import 'package:gruene_app/screens/search/search_screen.dart';
 import 'package:gruene_app/widget/scaffold_with_navbar.dart';
 
-import '../screens/start/start_screen.dart';
+import 'package:gruene_app/screens/start/start_screen.dart';
 
 final GlobalKey<NavigatorState> rootNavigatorKey = GlobalKey<NavigatorState>();
 final GlobalKey<NavigatorState> shellNavigatorKey = GlobalKey<NavigatorState>();
 
 final GoRouter router = GoRouter(
     navigatorKey: rootNavigatorKey,
-    initialLocation: routes['startScreen'] ?? '',
+    initialLocation: startScreen,
     routes: <RouteBase>[
       ShellRoute(
         navigatorKey: shellNavigatorKey,
@@ -24,13 +25,13 @@ final GoRouter router = GoRouter(
         routes: [
           // This screen is displayed on the ShellRoute's Navigator.
           GoRoute(
-            path: routes['startScreen']!,
+            path: startScreen,
             pageBuilder: (context, state) {
               return const NoTransitionPage(child: StartScreen());
             },
           ),
           GoRoute(
-            path: routes['searchScreen']!,
+            path: searchScreen,
             pageBuilder: (context, state) {
               return const NoTransitionPage(child: SearchScreen());
             },
@@ -39,7 +40,7 @@ final GoRouter router = GoRouter(
       ),
       GoRoute(
         parentNavigatorKey: rootNavigatorKey,
-        path: routes['news']!,
+        path: newsScreen,
         pageBuilder: (context, state) {
           return CustomTransitionPage(
             child: const NewsScreen(),
@@ -50,8 +51,6 @@ final GoRouter router = GoRouter(
         },
       ),
     ]);
-
-final routes = {'startScreen': '/', 'searchScreen': '/search', 'news': '/news'};
 
 SlideTransition slideAnimation(Animation<double> animation, Widget child) {
   const begin = Offset(1.0, 0.0);
