+++ lib/routing/routes.dart
@@ -1,0 +1,3 @@
+const startScreen = '/';
+const searchScreen = '/search';
+const newsScreen = '/news';
