+++ lib/screens/search/search_screen.dart
@@ -3,6 +3,7 @@ import 'package:flutter/src/widgets/framework.dart';
 import 'package:flutter/src/widgets/placeholder.dart';
 import 'package:go_router/go_router.dart';
 import 'package:gruene_app/routing/router.dart';
+import 'package:gruene_app/routing/routes.dart';
 
 class SearchScreen extends StatefulWidget {
   const SearchScreen({super.key});
@@ -19,8 +20,7 @@ class _SearchScreenState extends State<SearchScreen> {
         children: [
           Text('Search'),
           TextButton(
-              onPressed: () => context.push(routes['news']!),
-              child: Text('News'))
+              onPressed: () => context.push(newsScreen), child: Text('News'))
         ],
       ),
     );
