+++ lib/screens/start/start_screen.dart
@@ -1,5 +1,3 @@
-
-
 import 'package:flutter/material.dart';
 
 class StartScreen extends StatefulWidget {
@@ -10,8 +8,37 @@ class StartScreen extends StatefulWidget {
 }
 
 class _StartScreenState extends State<StartScreen> {
+  int _counter = 0;
+
+  void _incrementCounter() {
+    setState(() {
+      // This call to setState tells the Flutter framework that something has
+      // changed in this State, which causes it to rerun the build method below
+      // so that the display can reflect the updated values. If we changed
+      // _counter without calling setState(), then the build method would not be
+      // called again, and so nothing would appear to happen.
+      _counter++;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
-    return const Center(child: Text('StartScreen'));
+    return Column(
+      children: [
+        const Text(
+          'You have pushed the button this many times:',
+        ),
+        Text(
+          '$_counter',
+          style: Theme.of(context).textTheme.displayMedium,
+        ),
+        const Center(child: Text('StartScreen')),
+        FloatingActionButton(
+          onPressed: _incrementCounter,
+          tooltip: 'Increment',
+          child: const Icon(Icons.add),
+        ),
+      ],
+    );
   }
 }
